No need to set the public API define in an exe
Removing the call to set the export api define from MBProgram.

This was causing problems for filipp's Miracle-Grue unit test
program, the target name for which included a '/' or '\\'.

Rather than the quick hack to filter out the offending chars,
I figured we don't need to do this export on programs.