Add include paths pointing into Qt frameworks on mac
MW-2816

As part of the switch away from supporting 10.6 we're using the pre-built
Qt binaries. These only come as frameworks, and lack the include layout we
had been using until now. Because we no longer have those includes,
something needed to change with respect to finding headers.

We discussed three options:
* Adding the headers grabbed from elsewhere
* Changing all the #includes to use framework-style inclusion (<QtGui/QColor.h>)
* Adding an include path pointing into the framework

Of the three, changing the includes is the most 'correct'
We're going with the inclue paths into the frameworks because
it touches fewer things, and I can easily do it right now.