Adding a simple mustache builder/method pair
I'm doing stuff in argo that needs to use mustache, and
the existing method was just too complex for what I was
trying to do. This adds a tiny wrapper around mustache,
to just run it on a single file with a context and
write it to a given target.

It looks like some other places in the mustache gen
could be cleaned up a bit, but I'm not going to worry
about that right now.

Also, I left in a thing where I'm forcing utf8 all the
time. I probably need to make that an additional parameter.