Add a tool for running unit tests
MW-5426
http://jira.makerbot.net/browse/MW-5426

This includes some cross platform tools for adding paths to search for
dynamic libraries, and a tool to create virtual build targets for unit
tests.  Currently the only tool we have will run its test any time that
the "test" target is specified, regardless of whether any of the files
it is testing have actually changed.

I am not super confident whether the unit test dependency stuff works,
but so far the only place where we use this doesn't actually have
anything that needs to be built before running unit tests.